swagiaal: Changed DwarfDiegetDeclLine() to return int.
frysk-sys/lib/dwfl/ChangeLog
+2007-12-13  Sami Wagiaalla  <swagiaal@redhat.com>
+
+       * DwarfDie.java (getDeclLine): Now returns int instead of long.
+       * cni/DwarfDie.cxx (DwarfDie::get_decl_line): returns int
+       instead of long.
+

+++ frysk-sys/lib/dwfl/DwarfDie.java
@@ -95,7 +95,7 @@ abstract public class DwarfDie {
 	return get_decl_file(this.getPointer());
     }
   
-    public long getDeclLine() {
+    public int getDeclLine() {
 	return get_decl_line(this.getPointer());
     }
 
@@ -440,7 +440,7 @@ abstract public class DwarfDie {
   
     private native String get_decl_file (long var_die);
   
-    private native long get_decl_line (long var_die);
+    private native int get_decl_line (long var_die);
   
     private native int get_decl_column (long var_die);
